Update imports to ensure consistent schema usage
Previously, there were duplicates of the same means assessment schemas in crime-commons, the orchestrator was using a different version to the means assessment services, however, these have now been unified.

+++ maat-orchestration/src/test/java/uk/gov/justice/laa/crime/orchestration/mapper/MeansAssessmentMapperTest.java
@@ -136,7 +136,7 @@ class MeansAssessmentMapperTest {
         InitialAssessmentDTO initialAssessmentDTO = financialAssessmentDTO.getInitial();
         FullAssessmentDTO fullAssessmentDTO = financialAssessmentDTO.getFull();
         softly.assertThat(apiUpdateMeansAssessmentRequest.getFinancialAssessmentId())
-                .isEqualTo(BigDecimal.valueOf(financialAssessmentDTO.getId()));
+                .isEqualTo(financialAssessmentDTO.getId().intValue());
         softly.assertThat(apiUpdateMeansAssessmentRequest.getFullAssessmentDate())
                 .isEqualTo(toLocalDateTime(fullAssessmentDTO.getAssessmentDate()));
         softly.assertThat(apiUpdateMeansAssessmentRequest.getOtherHousingNote())
